fix: Rename store/index.ts to store/problems.ts
- Update corresponding import statements
- SetTotal.tsx: Add 'type' keyword for type import statement
